back-end/pyproject.toml
[tool.poetry]
name = "appstore-back-end"
version = "1.0.10"
description = ""
authors = ["Mathias and Tien Cheng"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
clearml = "^1.6.4"
fastapi = "^0.83.0"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
motor = "^3.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
click = "^8.1.3"
python-dotenv = "^0.21.0"
filetype = "^1.1.0"
password-strength = "^0.0.3.post2"
kubernetes = "^24.2.0"
Jinja2 = "^3.1.2"
fastapi-csrf-protect = "^0.2.2"
lxml = "^4.9.1"
bs4 = "^0.0.1"
minio = "^7.1.12"
colorama = "^0.4.6"
sse-starlette = "^1.2.1"
hypothesis = "^6.67.1"
miniopy-async = "^1.12"
aiohttp = "^3.8.4"

[tool.poetry.scripts]
boot = "main:start"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
mtools = "^1.7.0"
poethepoet = "^0.16.2"
pytest-asyncio = "^0.19.0"
pre-commit = "^2.20.0"
isort = "^5.10.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pylint = "^2.15.9"
pylint-pydantic = "^0.1.4"

[tool.poe.tasks.boot]
  help = "Boot application"
  cmd  = "uvicorn src.main:app --reload --host 0.0.0.0 --port 7070 --workers 4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.isort]
profile = "black"

[tool.pylint.MASTER]
load-plugins = "pylint_pydantic"
